compartmentalized login component in `Login.js`
