+++ app/package.json
@@ -26,6 +26,7 @@
     "react-dom": "^16.2.0",
     "react-native": "0.52.0",
     "react-native-elements": "^0.19.0",
-    "react-native-material-design": "^0.3.7"
+    "react-native-material-design": "^0.3.7",
+    "react-router-native": "^4.2.0"
   }
 }
